+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Notifications:
   - Add a new notification (disabled by default) to play a sound for each message.
   - The volume of each sound can now be controlled independently.
+  - The sound notification for each messages can be restricted to your own channel.
 
 ### 🐛 Bug Fix